[hotfix][kafka] Deduplicate reflection utils
