+++ flink-connectors/flink-connector-kafka/src/main/java/org/apache/flink/connector/kafka/sink/FlinkKafkaInternalProducer.java
@@ -30,10 +30,12 @@ import javax.annotation.Nullable;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
 import java.time.Duration;
 import java.util.Properties;
 
+import static org.apache.flink.connector.kafka.sink.ReflectionUtils.getField;
+import static org.apache.flink.connector.kafka.sink.ReflectionUtils.invoke;
+import static org.apache.flink.connector.kafka.sink.ReflectionUtils.setField;
 import static org.apache.flink.util.Preconditions.checkState;
 
 /**
@@ -177,7 +179,7 @@ class FlinkKafkaInternalProducer<K, V> extends KafkaProducer<K, V> {
     private void flushNewPartitions() {
         LOG.info("Flushing new partitions");
         TransactionalRequestResult result = enqueueNewPartitions();
-        Object sender = getField("sender");
+        Object sender = getProducerField("sender");
         invoke(sender, "wakeup");
         result.await();
     }
@@ -222,49 +224,8 @@ class FlinkKafkaInternalProducer<K, V> extends KafkaProducer<K, V> {
         }
     }
 
-    private static Object invoke(Object object, String methodName, Object... args) {
-        Class<?>[] argTypes = new Class[args.length];
-        for (int i = 0; i < args.length; i++) {
-            argTypes[i] = args[i].getClass();
-        }
-        return invoke(object, methodName, argTypes, args);
-    }
-
-    private static Object invoke(
-            Object object, String methodName, Class<?>[] argTypes, Object[] args) {
-        try {
-            Method method = object.getClass().getDeclaredMethod(methodName, argTypes);
-            method.setAccessible(true);
-            return method.invoke(object, args);
-        } catch (NoSuchMethodException | InvocationTargetException | IllegalAccessException e) {
-            throw new RuntimeException("Incompatible KafkaProducer version", e);
-        }
-    }
-
-    private Object getField(String fieldName) {
+    private Object getProducerField(String fieldName) {
         return getField(this, KafkaProducer.class, fieldName);
-    }
-
-    /**
-     * Gets and returns the field {@code fieldName} from the given Object {@code object} using
-     * reflection.
-     */
-    private static Object getField(Object object, String fieldName) {
-        return getField(object, object.getClass(), fieldName);
-    }
-
-    /**
-     * Gets and returns the field {@code fieldName} from the given Object {@code object} using
-     * reflection.
-     */
-    private static Object getField(Object object, Class<?> clazz, String fieldName) {
-        try {
-            Field field = clazz.getDeclaredField(fieldName);
-            field.setAccessible(true);
-            return field.get(object);
-        } catch (NoSuchFieldException | IllegalAccessException e) {
-            throw new RuntimeException("Incompatible KafkaProducer version", e);
-        }
     }
 
     /**
@@ -324,24 +285,6 @@ class FlinkKafkaInternalProducer<K, V> extends KafkaProducer<K, V> {
         }
     }
 
-    /**
-     * Sets the field {@code fieldName} on the given Object {@code object} to {@code value} using
-     * reflection.
-     */
-    private static void setField(Object object, String fieldName, Object value) {
-        setField(object, object.getClass(), fieldName, value);
-    }
-
-    private static void setField(Object object, Class<?> clazz, String fieldName, Object value) {
-        try {
-            Field field = clazz.getDeclaredField(fieldName);
-            field.setAccessible(true);
-            field.set(object, value);
-        } catch (NoSuchFieldException | IllegalAccessException e) {
-            throw new RuntimeException("Incompatible KafkaProducer version", e);
-        }
-    }
-
     @SuppressWarnings({"unchecked", "rawtypes"})
     private static Enum<?> getTransactionManagerState(String enumName) {
         try {
@@ -353,7 +296,7 @@ class FlinkKafkaInternalProducer<K, V> extends KafkaProducer<K, V> {
     }
 
     private Object getTransactionManager() {
-        return getField(TRANSACTION_MANAGER_FIELD_NAME);
+        return getProducerField(TRANSACTION_MANAGER_FIELD_NAME);
     }
 
     private static void transitionTransactionManagerStateTo(
